Fetches project clients for settings
Fetches and prepares the list of clients to be used in the project settings page.
This allows users to view and potentially update the client associated with a project.
The code currently fetches the clients and maps their display names, setting them to be used in a select component. The select component is currently commented out.